Remove unused files and simplify/refactor
