--- SimpleFeedReader/Models/INewsItemRepository.cs
@@ -1,11 +1,0 @@
-﻿using Microsoft.SyndicationFeed;
-using System;
-using System.Collections.Generic;
-
-namespace SimpleFeedReader.Models
-{
-    interface INewsItemRepository
-    {
-        List<ISyndicationItem> GetNewsItems(Uri feedUrl);
-    }
-}
