--- SimpleFeedReader/Pages/Contact.cshtml.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace SimpleFeedReader.Pages
-{
-    public class ContactModel : PageModel
-    {
-        public string Message { get; set; }
-
-        public void OnGet()
-        {
-            Message = "Your contact page.";
-        }
-    }
-}
