+++ SimpleFeedReader/Repositories/NewsItemRepository.cs
@@ -2,28 +2,31 @@
 using Microsoft.SyndicationFeed.Rss;
 using System;
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using System.Xml;
 
-namespace SimpleFeedReader.Models
+namespace SimpleFeedReader.Repositories
 {
-    public class NewsItemRepository : INewsItemRepository
+    public class NewsItemRepository
     {
-        public List<ISyndicationItem> GetNewsItems(Uri feedUrl)
+        public async Task<List<ISyndicationItem>> GetNewsItems(Uri feedUrl)
         {
             var returnList = new List<ISyndicationItem>();
 
-            using (var xmlReader = XmlReader.Create(feedUrl.ToString(), new XmlReaderSettings() { Async = true }))
+            using (var xmlReader = XmlReader.Create(feedUrl.ToString(), 
+                   new XmlReaderSettings { Async = true }))
             {
                 try
                 {
                     var feedReader = new RssFeedReader(xmlReader);
-                    while (feedReader.Read().Result)
+
+                    while (await feedReader.Read())
                     {
                         switch (feedReader.ElementType)
                         {
                             // RSS Item
                             case SyndicationElementType.Item:
-                                ISyndicationItem item = feedReader.ReadItem().Result;
+                                ISyndicationItem item = await feedReader.ReadItem();
                                 returnList.Add(item);
                                 break;
 
@@ -38,6 +41,7 @@ namespace SimpleFeedReader.Models
                     throw ae.Flatten();
                 }
             }
+
             return returnList;
         }
     }
