+++ SimpleFeedReader/Startup.cs
@@ -1,13 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.AspNetCore.Builder;
+﻿using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
-using Microsoft.EntityFrameworkCore;
-using SimpleFeedReader.Models;
+using SimpleFeedReader.Repositories;
 
 namespace SimpleFeedReader
 {
@@ -23,8 +18,8 @@ namespace SimpleFeedReader
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddScoped<NewsItemRepository>();
             services.AddMvc();
-
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
@@ -41,7 +36,6 @@ namespace SimpleFeedReader
             }
 
             app.UseStaticFiles();
-
             app.UseMvc();
         }
     }
